tror xml tingen skal være endret til floats

+++ BUILDER/src/main/java/com/game/ez_builder/builder/div/xmlBuilder.java
@@ -66,8 +66,8 @@ public class xmlBuilder {
     public void writeNewStatic(int heigth, int width, Vec2 position, int type, float rotation ,float radius, int texture){
         Element objectType = null;
 
-        int percent_h  = BuilderConstants.getScreenSize()/100;
-        int percent_w = BuilderConstants.getScreenSizeW()/100;
+        float percent_h  = BuilderConstants.getScreenSize()/100;
+        float percent_w = BuilderConstants.getScreenSizeW()/100;
 
         if(type == 1) {
             objectType = doc.createElement("BOX");
@@ -81,20 +81,20 @@ public class xmlBuilder {
         }
 
         Attr attr = doc.createAttribute("height");
-        attr.setValue("" + heigth/percent_h);
+        attr.setValue("" + (float)heigth/percent_h);
         objectType.setAttributeNode(attr);
 
 
         attr = doc.createAttribute("width");
-        attr.setValue(""+width/percent_w);
+        attr.setValue(""+(float)width/percent_w);
         objectType.setAttributeNode(attr);
 
        attr = doc.createAttribute("posX");
-        attr.setValue(""+(int)position.x/percent_w);
+        attr.setValue(""+position.x/percent_w);
         objectType.setAttributeNode(attr);
 
         attr = doc.createAttribute("posY");
-         attr.setValue(""+(int)position.y/percent_h);
+         attr.setValue(""+position.y/percent_h);
         objectType.setAttributeNode(attr);
 
         attr = doc.createAttribute("rotation");
